EFA transport: Add EFA transport support to Mooncake Transfer Engine

+++ mooncake-transfer-engine/include/transfer_metadata.h
@@ -52,8 +52,8 @@ class TransferMetadata {
         std::string name;
         uint64_t addr;
         uint64_t length;
-        std::vector<uint32_t> lkey;  // for rdma
-        std::vector<uint32_t> rkey;  // for rdma
+        std::vector<uint64_t> lkey;  // for rdma
+        std::vector<uint64_t> rkey;  // for rdma
         std::string shm_name;        // for nvlink and hip
         uint64_t offset;             // for cxl
     };
